requirements.txt

# Additional requirements go here

djangorestframework==3.7.7

django-filter==1.1.0
django-genomix==0.1.0
django-model-utils==3.0.0
